requirements.txt
attrs==25.3.0
Django==4.2.20
djangorestframework==3.15.2
drf-spectacular==0.28.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0


